+++ seqr/views/apis/summary_data_api_tests.py
@@ -6,7 +6,7 @@ import mock
 import responses
 
 from seqr.views.apis.summary_data_api import mme_details, success_story, saved_variants_page, hpo_summary_data, \
-    bulk_update_family_external_analysis, individual_metadata
+    bulk_update_family_external_analysis, individual_metadata, family_metadata, variant_metadata
 from seqr.views.utils.test_utils import AuthenticationTestCase, AnvilAuthenticationTestCase, AirtableTest, PARSED_VARIANTS
 from seqr.models import FamilyAnalysedBy, SavedVariant, VariantTag
 from settings import AIRTABLE_URL
@@ -701,12 +701,243 @@ class SummaryDataAPITest(AirtableTest):
         response = self.client.get(f'{gregor_projects_url}?includeAirtable=true')
         self._has_expected_metadata_response(response, multi_project_individuals, has_airtable=True, has_duplicate=True)
 
+    def test_family_metadata(self):
+        url = reverse(family_metadata, args=['R0003_test'])
+        self.check_require_login(url)
+        self.login_analyst_user()  # TODO test less access
+
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertListEqual(list(response_json.keys()), ['rows'])
+        self.assertListEqual(sorted([r['familyGuid'] for r in response_json['rows']]), ['F000011_11', 'F000012_12'])
+        test_row = next(r for r in response_json['rows'] if r['familyGuid'] == 'F000012_12')
+        self.assertDictEqual(test_row, {
+            'projectGuid': 'R0003_test',
+            'internal_project_id': 'Test Reprocessed Project',
+            'familyGuid': 'F000012_12',
+            'family_id': '12',
+            'displayName': '12',
+            'solve_status': 'Unsolved',
+            'actual_inheritance': 'unknown',
+            'date_data_generation': '2017-02-05',
+            'data_type': 'WES',
+            'proband_id': 'NA20889',
+            'maternal_id': '',
+            'paternal_id': '',
+            'other_individual_ids': 'NA20870; NA20888',
+            'individual_count': 3,
+            'family_structure': 'other',
+            'family_history': 'Yes',
+            'genes': 'DEL:chr1:249045487-249045898; OR4G11P',
+            'pmid_id': None,
+            'phenotype_description': None,
+            'analysisStatus': 'Q',
+            'analysis_groups': '',
+            'consanguinity': 'Unknown',
+        })
+
+        # Test all projects
+        all_projects_url = reverse(family_metadata, args=['all'])
+        response = self.client.get(all_projects_url)
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertListEqual(list(response_json.keys()), ['rows'])
+        self.assertListEqual(sorted([r['familyGuid'] for r in response_json['rows']]), [
+            'F000001_1', 'F000002_2', 'F000003_3', 'F000004_4', 'F000005_5', 'F000006_6', 'F000007_7', 'F000008_8',
+            'F000009_9', 'F000010_10', 'F000011_11', 'F000012_12', 'F000013_13'] + self.ADDITIONAL_FAMILIES)
+        test_row = next(r for r in response_json['rows'] if r['familyGuid'] == 'F000003_3')
+        self.assertDictEqual(test_row, {
+            'projectGuid': 'R0001_1kg',
+            'internal_project_id': '1kg project nåme with uniçøde',
+            'familyGuid': 'F000003_3',
+            'family_id': '3',
+            'displayName': '3',
+            'solve_status': 'Unsolved',
+            'actual_inheritance': '',
+            'date_data_generation': '2017-02-05',
+            'data_type': 'WES',
+            'other_individual_ids': 'NA20870',
+            'individual_count': 1,
+            'family_structure': 'singleton',
+            'genes': '',
+            'pmid_id': None,
+            'phenotype_description': None,
+            'analysisStatus': 'Q',
+            'analysis_groups': 'Accepted; Test Group 1',
+            'consanguinity': 'Unknown',
+            'condition_id': 'OMIM:615123',
+            'known_condition_name': '',
+            'condition_inheritance': 'Unknown',
+        })
+
+        # Test empty project
+        empty_project_url = reverse(family_metadata, args=['R0002_empty'])
+        response = self.client.get(empty_project_url)
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual(response.json(), {'rows': []})
+
+    def test_variant_metadata(self):
+        url = reverse(variant_metadata, args=[PROJECT_GUID])
+        self.check_require_login(url)
+        self.login_analyst_user()  # TODO test less access
+
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertListEqual(list(response_json.keys()), ['rows'])
+        row_ids = ['NA19675_1_21_3343353', 'HG00731_1_248367227', 'HG00731_19_1912634', 'HG00731_19_1912633', 'HG00731_19_1912632']
+        self.assertListEqual([r['genetic_findings_id'] for r in response_json['rows']], row_ids)
+        expected_row = {
+            **BASE_VARIANT_METADATA_ROW,
+            'additional_family_members_with_variant': 'HG00732',
+            'alt': 'T',
+            'chrom': '1',
+            'clinvar': {'alleleId': None, 'clinicalSignificance': '', 'goldStars': None, 'variationId': None},
+            'condition_id': 'MONDO:0044970',
+            'condition_inheritance': None,
+            'displayName': '2',
+            'familyGuid': 'F000002_2',
+            'family_id': '2',
+            'gene': 'RP11',
+            'gene_id': 'ENSG00000135953',
+            'gene_known_for_phenotype': 'Known',
+            'genetic_findings_id': 'HG00731_1_248367227',
+            'known_condition_name': 'mitochondrial disease',
+            'participant_id': 'HG00731',
+            'phenotype_contribution': 'Full',
+            'phenotype_description': 'microcephaly; seizures',
+            'pos': 248367227,
+            'projectGuid': 'R0001_1kg',
+            'internal_project_id': '1kg project nåme with uniçøde',
+            'ref': 'TC',
+            'tags': ['Known gene for phenotype'],
+            'variant_inheritance': 'paternal',
+            'variant_reference_assembly': 'GRCh37',
+            'zygosity': 'Homozygous',
+        }
+        self.assertDictEqual(response_json['rows'][1], expected_row)
+        expected_mnv = {
+            **BASE_VARIANT_METADATA_ROW,
+            'alt': 'T',
+            'chrom': '19',
+            'condition_id': 'MONDO:0044970',
+            'condition_inheritance': None,
+            'displayName': '2',
+            'end': 1912634,
+            'familyGuid': 'F000002_2',
+            'family_id': '2',
+            'gene': 'OR4G11P',
+            'gene_id': 'ENSG00000240361',
+            'gene_known_for_phenotype': 'Known',
+            'genetic_findings_id': 'HG00731_19_1912634',
+            'known_condition_name': 'mitochondrial disease',
+            'notes': 'The following variants are part of the multinucleotide variant 19-1912632-GC-TT (c.586_587delinsTT, p.Ala196Leu): 19-1912633-G-T, 19-1912634-C-T',
+            'participant_id': 'HG00731',
+            'phenotype_description': 'microcephaly; seizures',
+            'pos': 1912634,
+            'projectGuid': 'R0001_1kg',
+            'internal_project_id': '1kg project nåme with uniçøde',
+            'ref': 'C',
+            'tags': ['Known gene for phenotype'],
+            'transcript': 'ENST00000371839',
+            'variant_inheritance': 'unknown',
+            'variant_reference_assembly': 'GRCh38',
+            'zygosity': 'Heterozygous',
+        }
+        self.assertDictEqual(response_json['rows'][2], expected_mnv)
+
+        # Test gregor projects
+        gregor_projects_url = reverse(variant_metadata, args=['gregor'])
+        response = self.client.get(gregor_projects_url)
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertListEqual(list(response_json.keys()), ['rows'])
+        row_ids += ['NA20889_1_248367227', 'NA20889_1_249045487']
+        self.assertListEqual([r['genetic_findings_id'] for r in response_json['rows']], row_ids)
+        self.assertDictEqual(response_json['rows'][1], expected_row)
+        self.assertDictEqual(response_json['rows'][2], expected_mnv)
+        self.assertDictEqual(response_json['rows'][5], {
+            **BASE_VARIANT_METADATA_ROW,
+            'MME': True,
+            'alt': 'T',
+            'chrom': '1',
+            'clinvar': {'alleleId': None, 'clinicalSignificance': '', 'goldStars': None, 'variationId': None},
+            'condition_id': 'MONDO:0008788',
+            'displayName': '12',
+            'familyGuid': 'F000012_12',
+            'family_id': '12',
+            'family_history': 'Yes',
+            'gene': 'OR4G11P',
+            'gene_id': 'ENSG00000240361',
+            'gene_known_for_phenotype': 'Candidate',
+            'genetic_findings_id': 'NA20889_1_248367227',
+            'hgvsc': 'c.3955G>A',
+            'hgvsp': 'c.1586-17C>G',
+            'participant_id': 'NA20889',
+            'pos': 248367227,
+            'projectGuid': 'R0003_test',
+            'internal_project_id': 'Test Reprocessed Project',
+            'ref': 'TC',
+            'seqr_chosen_consequence': 'intron_variant',
+            'tags': ['Tier 1 - Novel gene and phenotype'],
+            'transcript': 'ENST00000505820',
+            'variant_inheritance': 'unknown',
+            'variant_reference_assembly': 'GRCh37',
+            'zygosity': 'Heterozygous',
+        })
+        self.assertDictEqual(response_json['rows'][6], {
+            **BASE_VARIANT_METADATA_ROW,
+            'alt': None,
+            'chrom': '1',
+            'condition_id': 'MONDO:0008788',
+            'displayName': '12',
+            'end': 249045898,
+            'familyGuid': 'F000012_12',
+            'family_id': '12',
+            'family_history': 'Yes',
+            'gene': None,
+            'gene_id': None,
+            'gene_known_for_phenotype': 'Candidate',
+            'genetic_findings_id': 'NA20889_1_249045487',
+            'participant_id': 'NA20889',
+            'pos': 249045487,
+            'projectGuid': 'R0003_test',
+            'internal_project_id': 'Test Reprocessed Project',
+            'ref': None,
+            'svType': 'DEL',
+            'sv_name': 'DEL:chr1:249045487-249045898',
+            'tags': ['Tier 1 - Novel gene and phenotype'],
+            'variant_inheritance': 'unknown',
+            'variant_reference_assembly': 'GRCh37',
+            'zygosity': 'Heterozygous',
+        })
+
+        # Test all projects
+        all_projects_url = reverse(variant_metadata, args=['all'])
+        response = self.client.get(all_projects_url)
+        self.assertEqual(response.status_code, 200)
+        response_json = response.json()
+        self.assertListEqual(list(response_json.keys()), ['rows'])
+        row_ids += self.ADDITIONAL_FINDINGS
+        self.assertListEqual([r['genetic_findings_id'] for r in response_json['rows']], row_ids)
+        self.assertDictEqual(response_json['rows'][1], expected_row)
+        self.assertDictEqual(response_json['rows'][2], expected_mnv)
+
+        # Test empty project
+        empty_project_url = reverse(family_metadata, args=['R0002_empty'])
+        response = self.client.get(empty_project_url)
+        self.assertEqual(response.status_code, 200)
+        self.assertDictEqual(response.json(), {'rows': []})
+
 
 # Tests for AnVIL access disabled
 class LocalSummaryDataAPITest(AuthenticationTestCase, SummaryDataAPITest):
     fixtures = ['users', '1kg_project', 'reference_data']
     NUM_MANAGER_SUBMISSIONS = 4
     ADDITIONAL_SAMPLES = ['NA21234', 'NA21987']
+    ADDITIONAL_FAMILIES = ['F000014_14']
+    ADDITIONAL_FINDINGS = ['NA21234_1_248367227']
 
 
 def assert_has_expected_calls(self, users, skip_group_call_idxs=None):
@@ -723,6 +954,8 @@ class AnvilSummaryDataAPITest(AnvilAuthenticationTestCase, SummaryDataAPITest):
     fixtures = ['users', 'social_auth', '1kg_project', 'reference_data']
     NUM_MANAGER_SUBMISSIONS = 4
     ADDITIONAL_SAMPLES = []
+    ADDITIONAL_FAMILIES = []
+    ADDITIONAL_FINDINGS = []
 
     def test_mme_details(self, *args):
         super(AnvilSummaryDataAPITest, self).test_mme_details(*args)
